% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{A stricter version of the pipe}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs, rhs}{the two sides of the pipe}
}
\description{
The drop-in replacement for the pipe operator will give a warning if they
input dataset is grouped, or the output dataset is grouped.
}
