% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{log_output}
\alias{log_output}
\title{Unified function for log output}
\usage{
log_output(text, message_type)
}
\arguments{
\item{text}{the text to output}

\item{message_type}{message type}
}
\description{
By default, all log output is via \code{stop()}
}
\details{
Set global environment variables to control the output behavour. Values must
be one of c('stop', 'warning', 'message', 'quiet').

e.g. \code{options(STRICTLYR_PIPE = 'message')}
}
